corrected drop

+++ main.py
@@ -211,8 +211,9 @@ def dropTables():
     cur.execute("SELECT table_schema,table_name FROM information_schema.tables WHERE table_schema = 'public' ORDER BY table_schema,table_name")
     rows = cur.fetchall()
     for row in rows:
-        print("dropping table: ", row[1])
-        cur.execute("drop table " + row[1] + " cascade")
+        if(row[1]!="pg_stat_statements"):
+            print("dropping table: ", row[1])
+            cur.execute("drop table " + row[1] + " cascade")
     conn.commit()
 
 def usarScripts():
